modules/s3-to-redshift-loader_3/requirements.txt
psycopg2-binary
redshift-auto-schema
fsspec
s3fs